Rename file to US spelling instead of english
