+++ src/utils/sanitizeDirectory.js
